feat(日志模块): 新增批量删除日志接口
1. 新增批量删除日志接口
2. 修改前端返回对象code类型为Integer

+++ blog-springboot/src/main/java/com/zrkizzy/blog/vo/Result.java
@@ -17,7 +17,7 @@ public class Result {
     /**
      * 状态码
      */
-    private long code;
+    private Integer code;
     /**
      * 提示信息
      */
